feat: expose File Write method
io/fs does not have any means of writting to fs.File for now, but since
Read already follows io.Reader (as expected) a Write method seems
reasonable and may be added to the official fs.File in the future. In
order to use this method a type assertion for gcsfs.File on the return of Open is
required.